add multi fields insert and retain pvc and vdc deployment

+++ workflow/base.py
@@ -1,5 +1,3 @@
-import time
-
 from deploy.configs.default_configs import DefaultConfigs
 from deploy.commons.common_params import CLUSTER, STANDALONE, Helm, Operator, DefaultRepository
 from deploy.client.default_client import DefaultClient
@@ -53,7 +51,7 @@ class Base:
 
         # Delete the service after the test is over
         if not param_info.deploy_retain and not param_info.deploy_skip:
-            self.set_teardown_funcs(self.deploy_delete)
+            self.set_teardown_funcs(self.deploy_delete, deploy_retain_pvc=param_info.deploy_retain_pvc)
 
         log.info("[setup_method] Test case: {0}, Test run_id: {1}".format(Report_Metric_Object.client.test_case_name,
                                                                           Report_Metric_Object.client.run_id))
@@ -85,16 +83,20 @@ class Base:
         elif len(_e) == 1:
             param_info.param_host = _e[0]
         else:
-            raise Exception("[Base] Can not parser endpoint: {0}, type: {1}, please check.".format(endpoint,
-                                                                                                   type(endpoint)))
+            raise Exception(f"[Base] Can not parser endpoint: {endpoint}, type: {type(endpoint)}, please check.")
 
     def init_server_client(self, deploy_tool=Operator, deploy_mode=STANDALONE):
         self.deploy_client = DefaultClient(deploy_tool=deploy_tool, deploy_mode=deploy_mode)
 
+    def set_global_function_before_test(self, release_name: str = ""):
+        # set global password
+        if release_name:
+            self.deploy_client.set_global_params(release_name=release_name)
+
     def deploy_default(self, deploy_tool=Operator, deploy_mode=STANDALONE, cpu=8, mem=16, other_config=None,
                        tag=None, repository=None, node_resources=None, set_dependence=None, input_configs: dict = {},
                        **kwargs):
-        tag = tag or param_info.milvus_tag or AutoGetTag().auto_tag()
+        tag = tag or param_info.milvus_tag or AutoGetTag().auto_tag(deploy_tool=deploy_tool)
         repository = repository or param_info.tag_repository
 
         # parser configs
@@ -103,7 +105,9 @@ class Base:
         # init server config
         config_obj = DefaultConfigs(deploy_tool=deploy_tool, deploy_mode=deploy_mode)
         custom_config = config_obj.setting_configs(node_resources=node_resources, set_dependence=set_dependence)
-        set_image = config_obj.set_image(tag=tag, repository=repository) if tag is not None else {}
+        set_image = config_obj.set_image(
+            tag=tag, repository=repository,
+            prefix=param_info.milvus_tag_prefix) if tag or param_info.milvus_tag_prefix else {}
         self.deploy_config = config_obj.server_resource(
             cpu=cpu, mem=mem, other_configs=[custom_config, input_configs, other_configs, set_image], **kwargs)
         log.info("[Base] deploy config: {}".format(self.deploy_config))
@@ -115,17 +119,18 @@ class Base:
         server_install_params = check_deploy_config(deploy_tool=deploy_tool, configs=self.deploy_config[0])
         self.deploy_release_name = self.deploy_client.install(server_install_params)
         self.deploy_client.wait_for_healthy(release_name=self.deploy_release_name)
-        endpoint = self.deploy_client.endpoint(release_name=self.deploy_release_name)
+        # endpoint = self.deploy_client.endpoint(release_name=self.deploy_release_name)
 
         # display server values
         log.debug(self.deploy_client.get_all_values(release_name=self.deploy_release_name))
         self.deploy_initial_state = self.deploy_client.get_pods(release_name=self.deploy_release_name)
 
         # set global host and port
-        self.parser_endpoint_to_global(endpoint)
+        # self.parser_endpoint_to_global(endpoint)
+        self.set_global_function_before_test(release_name=self.deploy_release_name)
 
         log.info("[Base] Service deployed successfully:{0}".format(self.deploy_release_name))
-        return self.deploy_release_name, endpoint
+        return self.deploy_release_name
 
     def upgrade_service(self, release_name=None, deploy_tool=Operator, deploy_mode=STANDALONE, upgrade_config=None):
         release_name = release_name or param_info.release_name
@@ -140,7 +145,8 @@ class Base:
         upgrade_configs = parser_input_config(input_content=upgrade_config)
 
         # format upgrade_config helm -> str, operator -> dict
-        format_upgrade_config = config_obj.config_conversion(upgrade_configs, update_helm_file=False)
+        format_upgrade_config = config_obj.config_conversion(upgrade_configs,
+                                                             update_helm_file=param_info.update_helm_file)
         self.upgrade_config = [format_upgrade_config, upgrade_configs]
 
         # check upgrade config and upgrade service
@@ -149,21 +155,23 @@ class Base:
         self.deploy_client.upgrade(server_upgrade_params)
 
         # wait healthy and get endpoint
-        wait_time_after_scale = 10
-        time.sleep(wait_time_after_scale)
+        # wait_time_after_scale = 10
+        # time.sleep(wait_time_after_scale)
         self.deploy_client.wait_for_healthy(release_name=release_name)
-        endpoint = self.deploy_client.endpoint(release_name=release_name)
+        # endpoint = self.deploy_client.endpoint(release_name=release_name)
 
         # display server values
         log.debug(self.deploy_client.get_all_values(release_name=release_name))
         self.deploy_initial_state = self.deploy_client.get_pods(release_name=release_name)
 
         # set global host and port
-        self.parser_endpoint_to_global(endpoint)
+        # self.parser_endpoint_to_global(endpoint)
+        self.set_global_function_before_test(release_name=self.deploy_release_name)
+
         log.info("[Base] Service upgraded successfully:{0}".format(self.deploy_release_name))
         return server_upgrade_params
 
-    def deploy_delete(self, deploy_client=None, deploy_release_name=""):
+    def deploy_delete(self, deploy_client=None, deploy_release_name="", deploy_retain_pvc=False, deploy_uninstall=True):
         deploy_client = deploy_client or self.deploy_client
         deploy_release_name = deploy_release_name or self.deploy_release_name or param_info.release_name
 
@@ -175,8 +183,10 @@ class Base:
 
             log.info("[Base] Start deleting services: {0}".format(deploy_release_name))
             deploy_client.get_pvc(release_name=deploy_release_name)
-            deploy_client.uninstall(release_name=deploy_release_name)
-            deploy_client.delete_pvc(release_name=deploy_release_name)
+            if deploy_uninstall:
+                deploy_client.uninstall(release_name=deploy_release_name)
+            if not deploy_retain_pvc:
+                deploy_client.delete_pvc(release_name=deploy_release_name)
             log.info("[Base] Service deleted successfully: {0}".format(deploy_release_name))
 
     def run_perf_case(self, callable_obj: callable, default_case_params, case_params, case_prepare, case_prepare_clean,
